lazai-integration/requirements.txt
alith>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
